+++ CHANGELOG.md
@@ -1,4 +1,8 @@
 # Changelog
 
+## Unreleased
+- `read` now supports datetimes, with new keyword arguments `convert_dates`,
+  `keep_default_dates` and `date_unit`.
+
 ## 1.0.0 (2023-02-07)
 - Initial release.